smart_file_organizer/requirements.txt
# Smart File Organizer - Dependencies
# Core System
watchdog>=4.0.0
pathlib>=1.0.1
pyyaml>=6.0

# Text Extraction
pytesseract>=0.3.10
pdf2image>=1.17.0
PyMuPDF>=1.24.0
python-docx>=1.1.0

# AI/ML Classification
ollama>=0.3.0
transformers>=4.40.0
torch>=2.0.0

# Security
cryptography>=42.0.0
pyzipper>=0.3.6
argon2-cffi>=23.1.0

# Deduplication
ImageHash>=4.3.1

# Utilities
python-magic-bin>=0.4.14; platform_system == "Windows"
python-magic>=0.4.27; platform_system != "Windows"
Send2Trash>=1.8.0
Pillow>=10.0.0
opencv-python>=4.9.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
